Move from Observable to Call
The issue I am trying to address is that some candidates feel pressured
to use rx even if they are absolutely unfamiliar with it.
In order to avoid setting them up for failure, we can just have a Call
by default.

If the candidate is familiar with Rx, we can easily encourage them
 to move back to Observable.